Fix style drift in tests
